requirements.txt
flask
flask-cors
flask-limiter
python-dotenv
requests

